Move SynchronizedDoujinClient to nhitomi/
