+++ nhitomi.Core/DoujinClientMethod.cs
@@ -1,0 +1,13 @@
+// Copyright (c) 2018-2019 fate/loli
+//
+// This software is released under the MIT License.
+// https://opensource.org/licenses/MIT
+
+namespace nhitomi.Core
+{
+    public enum DoujinClientMethod
+    {
+        Html,
+        Api
+    }
+}
