+++ nhitomi.Core/IDoujinClient.cs
@@ -27,68 +27,4 @@ namespace nhitomi.Core
         Task<IDoujin> GetAsync(string id, CancellationToken cancellationToken = default);
         Task<IAsyncEnumerable<IDoujin>> SearchAsync(string query, CancellationToken cancellationToken = default);
     }
-
-    public enum DoujinClientMethod
-    {
-        Html,
-        Api
-    }
-
-    public static class DoujinClientExtensions
-    {
-        sealed class SynchronizedClient : IDoujinClient
-        {
-            readonly SemaphoreSlim _semaphore = new SemaphoreSlim(1);
-            readonly IDoujinClient _impl;
-
-            public SynchronizedClient(IDoujinClient impl)
-            {
-                _impl = impl;
-            }
-
-            public string Name => _impl.Name;
-            public string Url => _impl.Url;
-            public string IconUrl => _impl.IconUrl;
-
-            public double RequestThrottle => _impl.RequestThrottle;
-
-            public DoujinClientMethod Method => _impl.Method;
-
-            public Regex GalleryRegex => _impl.GalleryRegex;
-
-            public async Task<IDoujin> GetAsync(string id, CancellationToken cancellationToken = default)
-            {
-                await _semaphore.WaitAsync(cancellationToken);
-                try
-                {
-                    return await _impl.GetAsync(id, cancellationToken);
-                }
-                finally
-                {
-                    _semaphore.Release();
-                }
-            }
-
-            public async Task<IAsyncEnumerable<IDoujin>> SearchAsync(
-                string query,
-                CancellationToken cancellationToken = default)
-            {
-                await _semaphore.WaitAsync(cancellationToken);
-                try
-                {
-                    return await _impl.SearchAsync(query, cancellationToken);
-                }
-                finally
-                {
-                    _semaphore.Release();
-                }
-            }
-
-            public override string ToString() => $"{nameof(SynchronizedClient)} ({_impl})";
-
-            public void Dispose() => _semaphore.Dispose();
-        }
-
-        public static IDoujinClient Synchronized(this IDoujinClient client) => new SynchronizedClient(client);
-    }
 }
